requirements.txt
fastapi
langchain
qdrant_client
python-dotenv